+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Compatibility:
 * Fix exception message when there are missing or extra keyword arguments - it contains all the missing/extra keywords now (#1522, @andrykonchin).
 * Always terminate native strings with enough `\0` bytes (#2704, @eregon).
 * Support `#dup` and `#clone` on foreign strings (@eregon).
+* Fix `Regexp.new` to coerce non-String arguments (#2705, @andrykonchin).
 
 Performance: